unify implementations across msbuild task and containerize exe
