+++ src/Containers/Microsoft.NET.Build.Containers/ContainerBuilder.cs
@@ -113,10 +113,15 @@ public static class ContainerBuilder
             // ports are validated by System.CommandLine API
             imageBuilder.ExposePort(number, type);
         }
+
         if (containerUser is { Length: > 0 } user)
         {
             imageBuilder.SetUser(user);
         }
+
+        ContainerHelpers.AssignUserFromEnvironment(imageBuilder, logger);
+        ContainerHelpers.AssignPortsFromEnvironment(imageBuilder, logger);
+
         BuiltImage builtImage = imageBuilder.Build();
         cancellationToken.ThrowIfCancellationRequested();
