+++ src/Containers/Microsoft.NET.Build.Containers/ImageBuilder.cs
@@ -1,10 +1,12 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.NET.Build.Containers;
-
+using System.Text;
 using System.Collections.ObjectModel;
 using Microsoft.NET.Build.Containers.Resources;
+
+namespace Microsoft.NET.Build.Containers;
+
 
 /// <summary>
 /// The class builds new image based on the base image.
